move src/js/chrome/ to /src/chrome
