+++ src/chrome/entry.ts
@@ -1,19 +1,18 @@
-import { globalFilterScope, registerModule, removeGlobalFilter, toggleGlobalFilter } from '../../redux/actions';
+import { globalFilterScope, registerModule, removeGlobalFilter, toggleGlobalFilter } from '../redux/actions';
 import { Listener } from '@redhat-cloud-services/frontend-components-utilities/MiddlewareListener';
-import { spinUpStore } from '../../redux/redux-config';
+import { spinUpStore } from '../redux/redux-config';
 import qe from './iqeEnablement';
-import consts from '../../utils/consts';
-import chromeHistory from '../../utils/chromeHistory';
-import { visibilityFunctions } from '../../utils/consts';
+import consts, { visibilityFunctions } from '../utils/consts';
+import chromeHistory from '../utils/chromeHistory';
 import Cookies from 'js-cookie';
-import { getEnv, getEnvDetails, getUrl, isBeta, isProd, updateDocumentTitle } from '../../utils/common';
+import { getEnv, getEnvDetails, getUrl, isBeta, isProd, updateDocumentTitle } from '../utils/common';
 import get from 'lodash/get';
-import { createSupportCase } from '../createCase';
-import * as actionTypes from '../../redux/action-types';
-import { flatTags } from '../App/GlobalFilter/globalFilterApi';
-import debugFunctions from '../debugFunctions';
-import { NavDOMEvent } from '../App/Sidenav/Navigation/ChromeLink';
-import { LibJWT } from '../auth';
+import { createSupportCase } from '../js/createCase';
+import * as actionTypes from '../redux/action-types';
+import { flatTags } from '../js/App/GlobalFilter/globalFilterApi';
+import debugFunctions from '../js/debugFunctions';
+import { NavDOMEvent } from '../js/App/Sidenav/Navigation/ChromeLink';
+import { LibJWT } from '../js/auth';
 import { ChromeAPI, ChromeUser } from '@redhat-cloud-services/types';
 
 type AppNavigationCB = (navEvent: { navId?: string; domEvent: NavDOMEvent }) => void;
