+++ src/chrome/initialize-jwt.test.js
