+++ src/chrome/initialize-jwt.ts
@@ -1,8 +1,8 @@
 import { ChromeUser } from '@redhat-cloud-services/types';
-import { LibJWT } from '../auth';
-import { decodeToken } from '../../jwt/jwt';
-import { spinUpStore } from '../../redux/redux-config';
-import { CacheAdapter } from '../../utils/cache';
+import { LibJWT } from '../js/auth';
+import { decodeToken } from '../jwt/jwt';
+import { spinUpStore } from '../redux/redux-config';
+import { CacheAdapter } from '../utils/cache';
 
 const initializeJWT = async (libjwt: LibJWT, chromeInstance: { cache?: CacheAdapter }) => {
   const { actions } = spinUpStore();
