+++ src/chrome/iqeEnablement.ts
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 /* eslint-disable prefer-rest-params */
 import type { Store } from 'redux';
-import { crossAccountBouncer } from '../auth';
-import { setGatewayError } from '../../redux/actions';
-import { get3scaleError } from '../../utils/responseInterceptors';
+import { crossAccountBouncer } from '../js/auth';
+import { setGatewayError } from '../redux/actions';
+import { get3scaleError } from '../utils/responseInterceptors';
 // TODO: Refactor this file to use modern JS
 
 let xhrResults: XMLHttpRequest[] = [];
